Tweak move error
Point at method definition that causes type to be consumed.

Fix #94056.
